chore(ProjectForm): modularize component
